updated from ' to "

+++ api/controllers/integrations.js
@@ -1,11 +1,11 @@
-const Integration = require('../models/integration');
+const Integration = require("../models/integration");
 
 exports.getApiKeys = async (req, res) => {
   try {
-    const apiKeys = await Integration.find({ owner: req.me._id }).populate('owner');
+    const apiKeys = await Integration.find({ owner: req.me._id }).populate("owner");
     res.status(200).json(apiKeys);
   } catch (error) {
-    console.log('error: ', error);
+    console.log("error: ", error);
     res.status(400);
   }
 }
@@ -14,17 +14,17 @@ exports.createApiKey = async (req, res) => {
     const apiKey = await Integration.create({ owner: req.me });
     res.status(201).json(apiKey);
   } catch (error) {
-    console.log('error: ', error);
+    console.log("error: ", error);
     res.status(400).send();
   }
 }
 exports.deleteApiKey = async (req, res) => {
   const { id } = req.params;
   try {
-    await Integration.findOneAndDelete({ _id: id });
+    await Integration.findOneAndDelete({ _id: id, owner: req.me._id });
     res.status(200).send();
   } catch (error) {
-    console.log('error: ', error);
+    console.log("error: ", error);
     res.status(400).send();
   }
 }
